src/main/steps/applicant/keep-details-private/private_details_not_confirmed/content.test.ts
import languageAssertions from '../../../../../test/unit/utils/languageAssertions';
import { FormContent } from '../../../../app/form/Form';
import { CommonContent } from '../../../common/common.content';
import { en as english, cy as welsh } from '../../../common/keep-details-private/private_details_not_confirmed/content';

import { generateContent } from './content';

const enContent = {
  ...english,
  section: 'Keeping your contact details private',
  continue: 'Save and continue',
};

const cyContent = {
  ...welsh,
  section: 'Cadw eich manylion cyswllt yn breifat',
  continue: 'Save and continue',
};

jest.mock('../../../../app/form/validation');
/* eslint-disable @typescript-eslint/ban-types */
describe('citizen-home content', () => {
  const commonContent = { language: 'en' } as CommonContent;
  let generatedContent;
  let form;
  beforeEach(() => {
    generatedContent = generateContent(commonContent);
    form = generatedContent.form as FormContent;
  });

  test('should return correct english content', () => {
    expect(generatedContent.section).toEqual('Keeping your contact details private');
    expect(generatedContent.title).toEqual('The court will not keep your contact details private');
    expect(generatedContent.line1).toEqual(
      'You have told us you do not want to keep your contact details private from the other people in this application.'
    );
  });

  // eslint-disable-next-line jest/expect-expect
  test('should return correct english content Data', () => {
    languageAssertions('en', enContent, () => generateContent(commonContent));
  });

  // eslint-disable-next-line jest/expect-expect
  test('should return correct welsh content', () => {
    languageAssertions('cy', cyContent, () => generateContent({ ...commonContent, language: 'cy' }));
  });

  test('should contain Save and continue button', () => {
    expect((form.submit?.text as Function)(generatedContent)).toBe('Save and continue');
  });
});
/* eslint-enable @typescript-eslint/ban-types */
